Consolidate ArbitraryInstances and arbitrary
