+++ tests/src/test/scala/org/http4s/CharsetRangeSuite.scala
@@ -17,7 +17,7 @@
 package org.http4s
 
 import cats.kernel.laws.discipline.EqTests
-import org.http4s.laws.discipline.ArbitraryInstances._
+import org.http4s.laws.discipline.arbitrary._
 import org.scalacheck.Arbitrary.arbitrary
 import org.scalacheck.Prop.{forAll, propBoolean}
