+++ tests/src/test/scala/org/http4s/HeadersSpec.scala
@@ -19,7 +19,7 @@ package org.http4s
 import cats.data.NonEmptyList
 import cats.kernel.laws.discipline.{MonoidTests, OrderTests}
 import org.http4s.headers._
-import org.http4s.laws.discipline.ArbitraryInstances._
+import org.http4s.laws.discipline.arbitrary._
 import org.http4s.syntax.header._
 import org.typelevel.ci._
