+++ tests/src/test/scala/org/http4s/MediaRangeSuite.scala
@@ -17,7 +17,7 @@
 package org.http4s
 
 import cats.kernel.laws.discipline.OrderTests
-import org.http4s.laws.discipline.ArbitraryInstances._
+import org.http4s.laws.discipline.arbitrary._
 import org.http4s.laws.discipline.HttpCodecTests
 
 class MediaRangeSuite extends Http4sSuite {
