+++ tests/src/test/scala/org/http4s/SchemeSuite.scala
@@ -21,7 +21,7 @@ import cats.kernel.laws.discipline.OrderTests
 import org.http4s.Uri.Scheme
 import org.http4s.internal.CharPredicate
 import org.http4s.laws.discipline.HttpCodecTests
-import org.http4s.laws.discipline.ArbitraryInstances._
+import org.http4s.laws.discipline.arbitrary._
 import org.http4s.util.Renderer
 import org.http4s.syntax.all._
 import org.scalacheck.Prop._
