+++ tests/src/test/scala/org/http4s/headers/AcceptCharsetSuite.scala
@@ -19,7 +19,7 @@ package headers
 
 import org.scalacheck.Prop._
 import org.http4s.syntax.all._
-import org.http4s.laws.discipline.ArbitraryInstances._
+import org.http4s.laws.discipline.arbitrary._
 
 class AcceptCharsetSuite extends HeaderLaws {
   checkAll("Accept-Charset", headerLaws[`Accept-Charset`])
