+++ tests/src/test/scala/org/http4s/headers/AcceptLanguageSuite.scala
@@ -18,7 +18,7 @@ package org.http4s
 package headers
 
 import org.http4s.syntax.all._
-import org.http4s.laws.discipline.ArbitraryInstances._
+import org.http4s.laws.discipline.arbitrary._
 import org.scalacheck.Prop._
 
 class AcceptLanguageSuite extends HeaderLaws {
