+++ tests/src/test/scala/org/http4s/headers/AgeSuite.scala
@@ -18,7 +18,7 @@ package org.http4s.headers
 
 import org.http4s.{ParseFailure, ParseResult}
 import org.http4s.syntax.header._
-import org.http4s.laws.discipline.ArbitraryInstances._
+import org.http4s.laws.discipline.arbitrary._
 import org.scalacheck.Prop._
 import scala.concurrent.duration._
