+++ tests/src/test/scala/org/http4s/headers/ContentEncodingSuite.scala
@@ -17,7 +17,7 @@
 package org.http4s
 package headers
 
-import org.http4s.laws.discipline.ArbitraryInstances._
+import org.http4s.laws.discipline.arbitrary._
 
 class ContentEncodingSuite extends HeaderLaws {
   checkAll("Content-Encoding", headerLaws[`Content-Encoding`])
