+++ tests/src/test/scala/org/http4s/headers/DateSuite.scala
@@ -18,7 +18,7 @@ package org.http4s
 package headers
 
 import java.time.{ZoneId, ZonedDateTime}
-import org.http4s.laws.discipline.ArbitraryInstances._
+import org.http4s.laws.discipline.arbitrary._
 import org.http4s.syntax.header._
 
 class DateSuite extends HeaderLaws {
