+++ tests/src/test/scala/org/http4s/headers/ForwardedSuite.scala
@@ -20,15 +20,12 @@ import java.nio.charset.StandardCharsets
 
 import cats.instances.string._
 import cats.syntax.option._
-import org.http4s.laws.discipline.ArbitraryInstances
+import org.http4s.laws.discipline.arbitrary._
 import org.http4s.{ParseFailure, Uri}
 import org.scalacheck.{Arbitrary, Gen}
 import org.scalacheck.Prop._
 
-class ForwardedSuite
-    extends munit.ScalaCheckSuite
-    with ArbitraryInstances
-    with ForwardedAuxiliaryGenerators {
+class ForwardedSuite extends munit.ScalaCheckSuite with ForwardedAuxiliaryGenerators {
 
   import Forwarded.Node
   import Node.{Name, Obfuscated}
