+++ tests/src/test/scala/org/http4s/headers/TransferEncodingSuite.scala
@@ -21,7 +21,7 @@ import cats.data.NonEmptyList
 import cats.syntax.foldable._
 import org.http4s.syntax.header._
 import org.scalacheck.Prop.forAll
-import org.http4s.laws.discipline.ArbitraryInstances._
+import org.http4s.laws.discipline.arbitrary._
 
 class TransferEncodingSuite extends HeaderLaws {
   checkAll("TransferEncoding", headerLaws[`Transfer-Encoding`])
